requirements.txt
Django==4.1
mysqlclient
marshmallow
bcrypt
pyjwt
requests
redis
celery
django-celery-results
django-celery-beat
djangorestframework
markdown       
django-filter
django-cors-headers